refactor: helper function for getting script hash from script purpose
This also moves the bits related to the script purpose/info into a
separate file

Fixes #1160
