+++ ledger/common/script/context.go
@@ -258,94 +258,6 @@ func (t TimeRange) ToPlutusData() data.PlutusData {
 	)
 }
 
-type ScriptInfo interface {
-	isScriptInfo()
-	ToPlutusData
-}
-
-type ScriptInfoMinting struct {
-	PolicyId lcommon.Blake2b224
-}
-
-func (ScriptInfoMinting) isScriptInfo() {}
-
-func (s ScriptInfoMinting) ToPlutusData() data.PlutusData {
-	return data.NewConstr(
-		0,
-		data.NewByteString(s.PolicyId.Bytes()),
-	)
-}
-
-type ScriptInfoSpending struct {
-	Input lcommon.TransactionInput
-	Datum data.PlutusData
-}
-
-func (ScriptInfoSpending) isScriptInfo() {}
-
-func (s ScriptInfoSpending) ToPlutusData() data.PlutusData {
-	if s.Datum == nil {
-		return data.NewConstr(
-			1,
-			s.Input.ToPlutusData(),
-		)
-	}
-	return data.NewConstr(
-		1,
-		s.Input.ToPlutusData(),
-		data.NewConstr(
-			0,
-			s.Datum,
-		),
-	)
-}
-
-type ScriptInfoRewarding struct {
-	StakeCredential lcommon.Credential
-}
-
-func (ScriptInfoRewarding) isScriptInfo() {}
-
-func (s ScriptInfoRewarding) ToPlutusData() data.PlutusData {
-	// TODO
-	return nil
-}
-
-type ScriptInfoCertifying struct {
-	Size        uint64
-	Certificate lcommon.Certificate
-}
-
-func (ScriptInfoCertifying) isScriptInfo() {}
-
-func (s ScriptInfoCertifying) ToPlutusData() data.PlutusData {
-	// TODO
-	return nil
-}
-
-type ScriptInfoVoting struct {
-	Voter lcommon.Voter
-}
-
-func (ScriptInfoVoting) isScriptInfo() {}
-
-func (s ScriptInfoVoting) ToPlutusData() data.PlutusData {
-	// TODO
-	return nil
-}
-
-type ScriptInfoProposing struct {
-	Size              uint64
-	ProposalProcedure lcommon.ProposalProcedure
-}
-
-func (ScriptInfoProposing) isScriptInfo() {}
-
-func (s ScriptInfoProposing) ToPlutusData() data.PlutusData {
-	// TODO
-	return nil
-}
-
 func sortInputs(inputs []lcommon.TransactionInput) []lcommon.TransactionInput {
 	ret := make([]lcommon.TransactionInput, len(inputs))
 	copy(ret, inputs)
@@ -483,67 +395,4 @@ func signatoriesInfo(
 		},
 	)
 	return tmp
-}
-
-type toScriptPurposeFunc func(lcommon.RedeemerKey) ScriptInfo
-
-// scriptPurposeBuilder creates a reusable function preloaded with information about a particular transaction
-func scriptPurposeBuilder(
-	resolvedInputs []lcommon.Utxo,
-	inputs []lcommon.TransactionInput,
-	mint lcommon.MultiAsset[lcommon.MultiAssetTypeMint],
-	// TODO: certificates
-	withdrawals map[*lcommon.Address]uint64,
-	// TODO: proposal procedures
-	// TODO: votes
-) toScriptPurposeFunc {
-	return func(redeemerKey lcommon.RedeemerKey) ScriptInfo {
-		// TODO: implement additional redeemer tags
-		// https://github.com/aiken-lang/aiken/blob/af4e04b91e54dbba3340de03fc9e65a90f24a93b/crates/uplc/src/tx/script_context.rs#L771-L826
-		switch redeemerKey.Tag {
-		case lcommon.RedeemerTagSpend:
-			var datum data.PlutusData
-			tmpInput := inputs[redeemerKey.Index]
-			for _, resolvedInput := range resolvedInputs {
-				if resolvedInput.Id.String() == tmpInput.String() {
-					if tmpDatum := resolvedInput.Output.Datum(); tmpDatum != nil {
-						datum = tmpDatum.Data
-					}
-					break
-				}
-			}
-			return ScriptInfoSpending{
-				Input: inputs[redeemerKey.Index],
-				Datum: datum,
-			}
-		case lcommon.RedeemerTagMint:
-			// TODO: fix this to work for more than one minted policy
-			mintPolicies := mint.Policies()
-			slices.SortFunc(
-				mintPolicies,
-				func(a, b lcommon.Blake2b224) int { return bytes.Compare(a.Bytes(), b.Bytes()) },
-			)
-			return ScriptInfoMinting{
-				PolicyId: mintPolicies[redeemerKey.Index],
-			}
-		case lcommon.RedeemerTagCert:
-			return nil
-		case lcommon.RedeemerTagReward:
-			return nil
-		case lcommon.RedeemerTagVoting:
-			return nil
-		case lcommon.RedeemerTagProposing:
-			return nil
-		}
-		return nil
-	}
-}
-
-func scriptPurposeStripDatum(purpose ScriptInfo) ScriptInfo {
-	switch p := purpose.(type) {
-	case ScriptInfoSpending:
-		p.Datum = nil
-		return p
-	}
-	return purpose
 }
